Major refactoring of gen_tile_config()
First of all, readability is improved a lot by reducing branches and
if-clauses, partially achieved by migrating the code assembling
filenames and coordinates to the MosaicData class.

Additionally, an optional parameter 'force_2d' is introduced to request
the resulting tile configuration to be for 2D datasets even if the
original mosaic consists of 3D stacks. This is intended for e.g.
stitching projections or single slices of the original dataset.